# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MBS CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads/{id}/drip:
    get:
      operationId: getLeadDripEnrollment
      tags: [drip]
      summary: Get current drip enrollment for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Current enrollment or null
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/DripEnrollment"
                  - type: "null"
  /leads/{id}/drip/enroll:
    post:
      operationId: enrollLeadInDrip
      tags: [drip]
      summary: Manually enroll a lead in a drip sequence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sequenceId]
              properties:
                sequenceId:
                  type: integer
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DripEnrollment"
  /leads/{id}/drip/unenroll:
    post:
      operationId: unenrollLeadFromDrip
      tags: [drip]
      summary: Unenroll a lead from their active drip sequence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unenrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DripEnrollment"
  /drip/sequences:
    get:
      operationId: listDripSequences
      tags: [drip]
      summary: List all drip sequences
      responses:
        "200":
          description: Sequences list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DripSequence"
    post:
      operationId: createDripSequence
      tags: [drip]
      summary: Create a drip sequence (all approved users; reps own the result)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DripSequenceInput"
      responses:
        "201":
          description: Created sequence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DripSequence"
  /drip/sequences/{id}:
    get:
      operationId: getDripSequence
      tags: [drip]
      summary: Get a drip sequence with its steps
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sequence with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DripSequenceDetail"
    put:
      operationId: updateDripSequence
      tags: [drip]
      summary: Update a drip sequence (managers/admins or owning rep)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DripSequenceInput"
      responses:
        "200":
          description: Updated sequence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DripSequence"
    delete:
      operationId: deleteDripSequence
      tags: [drip]
      summary: Delete a drip sequence (reps may delete only their own)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Sequence deleted
  /drip/sequences/{id}/steps:
    put:
      operationId: upsertDripSequenceSteps
      tags: [drip]
      summary: Replace all steps of a drip sequence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [steps]
              properties:
                steps:
                  type: array
                  items:
                    $ref: "#/components/schemas/DripStepInput"
      responses:
        "200":
          description: Updated steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DripSequenceStep"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DripSequence:
      type: object
      required: [id, name, triggerStatus, isActive, stepCount, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        triggerStatus:
          type: string
        senderMode:
          type: string
          enum: [template, default, assigned_rep]
        isActive:
          type: boolean
        stepCount:
          type: integer
        createdBy:
          type: ["integer", "null"]
        creator:
          oneOf:
            - type: object
              properties:
                id:
                  type: integer
                name:
                  type: ["string", "null"]
                email:
                  type: string
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DripSequenceDetail:
      allOf:
        - $ref: "#/components/schemas/DripSequence"
        - type: object
          properties:
            steps:
              type: array
              items:
                $ref: "#/components/schemas/DripSequenceStep"
    DripSequenceInput:
      type: object
      required: [name, triggerStatus]
      properties:
        name:
          type: string
        triggerStatus:
          type: string
        senderMode:
          type: string
          enum: [template, default, assigned_rep]
        isActive:
          type: boolean
    DripSequenceStep:
      type: object
      required: [id, sequenceId, stepOrder, templateId, delayHours, createdAt]
      properties:
        id:
          type: integer
        sequenceId:
          type: integer
        stepOrder:
          type: integer
        templateId:
          type: integer
        template:
          oneOf:
            - type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
                subject:
                  type: string
            - type: "null"
        delayHours:
          type: integer
        createdAt:
          type: string
          format: date-time
    DripStepInput:
      type: object
      required: [templateId, delayHours]
      properties:
        templateId:
          type: integer
        delayHours:
          type: integer
    DripEnrollment:
      type: object
      required: [id, leadId, sequenceId, currentStep, status, enrolledAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        sequenceId:
          type: integer
        sequence:
          oneOf:
            - type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
                steps:
                  type: integer
            - type: "null"
        currentStep:
          type: integer
        status:
          type: string
          enum: [active, completed, unenrolled]
        enrolledAt:
          type: string
          format: date-time
        lastStepSentAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        unenrolledAt:
          type: ["string", "null"]
          format: date-time
...